+++ test/urlencoded.js
@@ -587,9 +587,11 @@ describe('bodyParser.urlencoded()', function () {
     })
 
     it('should error from verify', function (done) {
-      var server = createServer({ verify: function (req, res, buf) {
-        if (buf[0] === 0x20) throw new Error('no leading space')
-      } })
+      var server = createServer({
+        verify: function (req, res, buf) {
+          if (buf[0] === 0x20) throw new Error('no leading space')
+        }
+      })
 
       request(server)
         .post('/')
@@ -599,9 +601,11 @@ describe('bodyParser.urlencoded()', function () {
     })
 
     it('should error with type = "entity.verify.failed"', function (done) {
-      var server = createServer({ verify: function (req, res, buf) {
-        if (buf[0] === 0x20) throw new Error('no leading space')
-      } })
+      var server = createServer({
+        verify: function (req, res, buf) {
+          if (buf[0] === 0x20) throw new Error('no leading space')
+        }
+      })
 
       request(server)
         .post('/')
@@ -612,12 +616,14 @@ describe('bodyParser.urlencoded()', function () {
     })
 
     it('should allow custom codes', function (done) {
-      var server = createServer({ verify: function (req, res, buf) {
-        if (buf[0] !== 0x20) return
-        var err = new Error('no leading space')
-        err.status = 400
-        throw err
-      } })
+      var server = createServer({
+        verify: function (req, res, buf) {
+          if (buf[0] !== 0x20) return
+          var err = new Error('no leading space')
+          err.status = 400
+          throw err
+        }
+      })
 
       request(server)
         .post('/')
@@ -627,12 +633,14 @@ describe('bodyParser.urlencoded()', function () {
     })
 
     it('should allow custom type', function (done) {
-      var server = createServer({ verify: function (req, res, buf) {
-        if (buf[0] !== 0x20) return
-        var err = new Error('no leading space')
-        err.type = 'foo.bar'
-        throw err
-      } })
+      var server = createServer({
+        verify: function (req, res, buf) {
+          if (buf[0] !== 0x20) return
+          var err = new Error('no leading space')
+          err.type = 'foo.bar'
+          throw err
+        }
+      })
 
       request(server)
         .post('/')
@@ -643,9 +651,11 @@ describe('bodyParser.urlencoded()', function () {
     })
 
     it('should allow pass-through', function (done) {
-      var server = createServer({ verify: function (req, res, buf) {
-        if (buf[0] === 0x5b) throw new Error('no arrays')
-      } })
+      var server = createServer({
+        verify: function (req, res, buf) {
+          if (buf[0] === 0x5b) throw new Error('no arrays')
+        }
+      })
 
       request(server)
         .post('/')
@@ -655,9 +665,11 @@ describe('bodyParser.urlencoded()', function () {
     })
 
     it('should 415 on unknown charset prior to verify', function (done) {
-      var server = createServer({ verify: function (req, res, buf) {
-        throw new Error('unexpected verify call')
-      } })
+      var server = createServer({
+        verify: function (req, res, buf) {
+          throw new Error('unexpected verify call')
+        }
+      })
 
       var test = request(server).post('/')
       test.set('Content-Type', 'application/x-www-form-urlencoded; charset=x-bogus')