+++ HISTORY.md
@@ -3,7 +3,7 @@ unreleased
 
   * deps: http-errors@1.7.3
     - deps: inherits@2.0.4
-  * deps: qs@6.9.1
+  * deps: qs@6.9.3
   * deps: raw-body@2.4.1
     - deps: http-errors@1.7.3
   * deps: safe-buffer@5.2.0