+++ HISTORY.md
@@ -6,7 +6,7 @@ unreleased
   * deps: qs@6.9.3
   * deps: raw-body@2.4.1
     - deps: http-errors@1.7.3
-  * deps: safe-buffer@5.2.0
+  * deps: safe-buffer@5.2.1
   * deps: type-is@~1.6.18
 
 1.19.0 / 2019-04-25