+++ HISTORY.md
@@ -4,8 +4,9 @@ unreleased
   * deps: http-errors@1.7.3
     - deps: inherits@2.0.4
   * deps: qs@6.9.6
-  * deps: raw-body@2.4.1
-    - deps: http-errors@1.7.3
+  * deps: raw-body@2.4.2
+    - deps: bytes@3.1.1
+    - deps: http-errors@1.8.1
   * deps: safe-buffer@5.2.1
   * deps: type-is@~1.6.18